requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
decorator==4.4.2
idna==2.10
pycparser==2.20
python-telegram-bot==13.1
pytz==2020.5
requests==2.25.1
six==1.15.0
tornado==6.1
tqdm==4.54.1
tzlocal==2.1
urllib3==1.26.2
